Removed obsolete fl.datasets.lightning.DataModule
