+++ src/fl/configs/mlp.yaml
@@ -11,17 +11,17 @@ node:
   model:
     name: mlp_regressor
     batch_size: 256
-    l1: 1e-3
+    l1: 2e-4
     hidden_size: 2048
   optimizer:
     name: sgd
     lr: 1e-4
-    weight_decay: 1e-2
+    weight_decay: 1e-3
   #scheduler:
   #  name: ExponentialLR
   scheduler:
     name: exponential_lr
-    rounds: 32
+    rounds: 24
     epochs_in_round: 4
     gamma: 0.95
   training:
@@ -55,7 +55,7 @@ dataset:
   gwas: '${dataset.split.path}/${dataset.phenotype.name}/node_${node.index}/fold_${fold.index}.gwas.tsv'
 
 server:
-  rounds: 32
+  rounds: 24
   checkpoint_dir: 'models/${dataset.phenotype.name}/fold_${fold.index}'
   strategy:
     name: fedavg
