+++ src/nn/lightning.py
@@ -17,19 +17,19 @@ class DataModule(LightningDataModule):
         self.batch_size = batch_size
 
     def train_dataloader(self) -> DataLoader:
-        loader = DataLoader(self.train_dataset, batch_size=self.batch_size, shuffle=True, num_workers=2)
+        loader = DataLoader(self.train_dataset, batch_size=self.batch_size, shuffle=True, num_workers=1)
         return loader
     
     def val_dataloader(self) -> DataLoader:
-        loader = DataLoader(self.val_dataset, batch_size=self.batch_size, shuffle=False, num_workers=2)
+        loader = DataLoader(self.val_dataset, batch_size=self.batch_size, shuffle=False, num_workers=1)
         return loader
     
     def test_dataloader(self) -> DataLoader:
-        loader = DataLoader(self.test_dataset, batch_size=self.batch_size, shuffle=False, num_workers=2)
+        loader = DataLoader(self.test_dataset, batch_size=self.batch_size, shuffle=False, num_workers=1)
         return loader
 
     def predict_dataloader(self) -> List[DataLoader]:
-        train_loader = DataLoader(self.train_dataset, batch_size=self.batch_size, shuffle=False, num_workers=2)
+        train_loader = DataLoader(self.train_dataset, batch_size=self.batch_size, shuffle=False, num_workers=1)
         val_loader = self.val_dataloader()
         test_loader = self.test_dataloader()
         return [train_loader, val_loader, test_loader]
@@ -45,3 +45,6 @@ class DataModule(LightningDataModule):
     
     def test_len(self):
         return len(self.test_dataset)
+
+    def feature_count(self):
+        return self._X_train.shape[1]
